pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolo-from-scratch"
version = "0.1.0"
description = "YOLO implementation from scratch"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "opencv-python",
    "pillow",
    "pyyaml",
    "matplotlib",
    "tqdm",
    "torch",
    "torchvision",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }